Move fuzz helper code into a separate package
